efi: Add support for UEFI v2.10 GetMemoryMapEx
This extension for GetMemoryMap allows the guest OS to declare its
support for a particular memory feature that may have been added in
a later UEFI spec. In particular, this is useful to declare support
for the unaccepted memory type, so that the UEFI is free to return
a memory map with that memory type instead of accept all the memory
before continuing boot.

CC: Thomas Lendacky <thomas.lendacky@amd.com>
CC: Kirill Shutemov <kirill.shutemov@intel.com>
CC: Jiewen Yao <jiewen.yao@intel.com>
CC: Gerd Hoffman <kraxel@redhat.com>

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>