x86/boot: Add boot_features to kernel_info
The addition of unaccepted memory to the UEFI v2.9 specification
means that the UEFI is allowed to create memory regions that (a
confidential VM) has not yet accepted into the guest. If the OS does not
know about unaccepted memory regions, then it can be stuck with
far less system memory than expected.

By adding this boot feature information to an easily-accessible location
in the kernel binary, the bootloader can assess if it should keep
the unaccepted memory unaccepted, or request that the UEFI accepts all
memory before continuing.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>