efi: Add OsIndications variable set on boot
For unaccepted memory to be used in the firmware, the firmware must know
that the OS will understand it. That information is communicated through
the UEFI standard OsIndications variable. We only set this bit if the
OsIndicationsSupported bitmask also has it set, otherwise we will cause
an unformalized OsIndications variable error.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>