x86/boot: Add unaccepted memory kernel_info
The addition of UEFI v2.9's memory type EfiUnacceptedMemory brings with
it the predicament of whether the firmware should create unaccepted
memory for a kernel that isn't built with CONFIG_UNACCEPTED_MEMORY=y or
if it should accept all memory before ExitBootServices().

Since unaccepted memory is configurable, it is up to the bootloader to
know whether the kernel will be able to handle unaccepted memory. By
providing this kernel_info variable length entry, we give a firm
positive or negative signal whether boot services can be ended before
accepting all memory.

Cc: Thomas Lendacky <Thomas.Lendacky@amd.com>
Cc: "Min M., Xu" <min.m.xu@intel.com>
Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Kirill Shutemov <kirill.shutemov@intel.com>
Cc: Ashish Kalra <Ashish.Kalra@amd.com>
Cc: Ard Biesheuvel <ardb@kernel.org>
Cc: Paolo Bonzini <pbonzini@redhat.org>

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>