KVM: SVM: move SEV fd access check to ioctl
All ccp API calls ought to be under presupposed permission to use the
driver. Given that sev_do_cmd is already external, the
permission-checking falls to the API user.

Check for ccp device access at the top of the mem_enc_ioctl once the
kvm_cmd_arg is copied from user space to not have to check any deeper
in the call stack.

Change-Id: Id9696c524dc4e29c04a0bd4094b77826c95e6892
Signed-off-by: Dionna Glaze <dionnaglaze@google.com>