drivers/tee: Add SVSM-TEE support
A Secure VM Services Module is an in-VM TEE that is able to provide a
secure implementation of a TPM. Add a relatively basic TEE device for
invoking vTPM commands when using a session the fTPM trusted
application.

Generic SVSM-as-TEE support for invoking any SVSM protocol is not a
goal, just fTPM support.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>