virt/coco/tsm: Add manifest_selector WO binary attribute
For service providers where a single service may need to attest a high
cardinality space of possibilities, a single manifest version is
impractical, and the service needs to be told what specifically is
being requested to attest.

The service_provider is responsible for providing a cryptographic binding
between the inblob, the service_guid, the manifest_selector, the
manifestblob, and the manifest_version in the resulting outblob.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>