virt/coco/tsm: Add service_error RO attribute
A service provider may return an error code different from -EINVAL when
requesting a manifestblob, so make that result code readable.

Given that different services may report errors differently, add a new
provider operation, service_error_show, which will be called when the
attribute is visible, is requested to be shown, and the service error
is non-NULL.

A NULL value for the service error results in an empty service_error
attribute.

The AMD SVSM attestation protocol specifies protocol-specific result
codes to be within the 0x8000_1000-0xFFFF_FFFF range.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>